pyproject.toml
[project]
name = "task-email"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.1",
    "openai-agents>=0.7.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
]
